[refact] move actor/critic and others inside workers
