+++ examples/ppo_trainer/run_deepseek_full_hh_rlhf.sh
@@ -10,7 +10,7 @@ python3 -m verl.trainer.main_ppo --config-path=./config --config-name='ppo_megat
     data.val_batch_size=128 \
     data.max_prompt_length=128 \
     data.max_response_length=128 \
-    actor_rollout_ref.model.path=deepseek-ai/deepseek-llm-7b-chat \
+    actor_rollout_ref.model.path=hdfs://haruna/home/byte_data_seed/lf_lq/user/zhangchi.usc1992/models/deepseek-llm-7b-chat \
     actor_rollout_ref.actor.optim.lr=1e-6 \
     actor_rollout_ref.actor.ppo_mini_batch_size=128 \
     actor_rollout_ref.actor.ppo_micro_batch_size=16 \
@@ -21,7 +21,7 @@ python3 -m verl.trainer.main_ppo --config-path=./config --config-name='ppo_megat
     actor_rollout_ref.ref.log_prob_micro_batch_size=16 \
     actor_rollout_ref.ref.param_offload=False \
     critic.optim.lr=1e-5 \
-    critic.model.path=deepseek-ai/deepseek-llm-7b-chat \
+    critic.model.path=hdfs://haruna/home/byte_data_seed/lf_lq/user/zhangchi.usc1992/models/deepseek-llm-7b-chat \
     critic.model.enable_gradient_checkpointing=False \
     critic.ppo_micro_batch_size=16 \
     reward_model.enable=True \
