+++ examples/ppo_trainer/run_deepseek_megatron.sh
@@ -1,13 +1,13 @@
 set -x
 
 python3 -m verl.trainer.main_ppo --config-path=./config --config-name='ppo_megatron_trainer'\
-    data.train_files=$HOME/data/gsm8k/train.parquet \
-    data.val_files=$HOME/data/gsm8k/test.parquet \
+    data.train_files=hdfs://haruna/home/byte_data_seed/lf_lq/user/zhangchi.usc1992/data/rlhf/gsm8k/train.parquet \
+    data.val_files=hdfs://haruna/home/byte_data_seed/lf_lq/user/zhangchi.usc1992/data/rlhf/gsm8k/test.parquet \
     data.train_batch_size=1024 \
     data.val_batch_size=1312 \
     data.max_prompt_length=512 \
     data.max_response_length=512 \
-    actor_rollout_ref.model.path=deepseek-ai/deepseek-coder-6.7b-instruct \
+    actor_rollout_ref.model.path=hdfs://haruna/home/byte_data_seed/lf_lq/user/zhangchi.usc1992/experiments/gsm8k/deepseek-coder-6.7b-instruct/global_step_58 \
     actor_rollout_ref.actor.optim.lr=2e-6 \
     actor_rollout_ref.actor.ppo_mini_batch_size=256 \
     actor_rollout_ref.actor.ppo_micro_batch_size=64 \
@@ -17,7 +17,7 @@ python3 -m verl.trainer.main_ppo --config-path=./config --config-name='ppo_megat
     actor_rollout_ref.rollout.gpu_memory_utilization=0.5 \
     actor_rollout_ref.ref.log_prob_micro_batch_size=128 \
     critic.optim.lr=2e-5 \
-    critic.model.path=deepseek-ai/deepseek-coder-6.7b-instruct \
+    critic.model.path=hdfs://haruna/home/byte_data_seed/lf_lq/user/zhangchi.usc1992/experiments/gsm8k/deepseek-coder-6.7b-instruct/global_step_58 \
     critic.model.enable_gradient_checkpointing=False \
     critic.ppo_micro_batch_size=64 \
     algorithm.kl_ctrl.kl_coef=0.001 \
